+++ core/chain/tx/broadcast/resolvers/cosmos.ts
@@ -3,7 +3,7 @@ import { getCosmosClient } from '@core/chain/chains/cosmos/client'
 import { attempt } from '@lib/utils/attempt'
 import { isInError } from '@lib/utils/error/isInError'
 
-import { BroadcastTxResolver } from './BroadcastTxResolver'
+import { BroadcastTxResolver } from '../resolver'
 
 export const broadcastCosmosTx: BroadcastTxResolver<CosmosChain> = async ({
   chain,
