+++ core/chain/tx/broadcast/resolvers/tron.ts
@@ -1,8 +1,8 @@
 import { OtherChain } from '@core/chain/Chain'
 import { queryUrl } from '@lib/utils/query/queryUrl'
 
-import { tronRpcUrl } from '../../chains/tron/config'
-import { BroadcastTxResolver } from './BroadcastTxResolver'
+import { tronRpcUrl } from '../../../chains/tron/config'
+import { BroadcastTxResolver } from '../resolver'
 
 export const broadcastTronTx: BroadcastTxResolver<OtherChain.Tron> = async ({
   tx,
